Fixed testing when running all tests: Removing handlers by hand did the trick. Renamed test files to lower case to make it unittests run on vscode.

+++ Tests/test_training.py
@@ -17,11 +17,11 @@ class TestTraining(unittest.TestCase):
         self.training_data_paths = [resources.test_training_src_dir / '2019-07-11_15-14-48_100p']
         self.load_datasets_path = resources.test_training_src_dir / '2019-09-06_15-44-58_63_sensors'
         self.expected_num_epochs_during_training = 1
-
-    def test_training(self):
         self.st = SensorTrainer(data_source_paths=self.training_data_paths,
                            save_datasets_path=self.training_save_path,
                            epochs=self.expected_num_epochs_during_training)
+
+    def test_training(self):
         self.st.run_training()
         dirs = [e for e in self.training_save_path.iterdir() if e.is_dir()]
         with open(dirs[0] / 'output.log') as f:
@@ -31,7 +31,11 @@ class TestTraining(unittest.TestCase):
 
     def tearDown(self) -> None:
         self.st.training_data_generator.end_threads()
-        shutil.rmtree(self.training_save_path)
+        logging.shutdown()
+        r = logging.getLogger("")
+        [r.removeHandler(x) for x in r.handlers]
+        if self.training_save_path.exists():
+            shutil.rmtree(self.training_save_path)
         logging.shutdown()
 
 
@@ -73,6 +77,9 @@ class TestTrainingWithFixedDataset(unittest.TestCase):
 
     def tearDown(self) -> None:
         self.gen.end_threads()
+        logging.shutdown()
+        r = logging.getLogger("")
+        [r.removeHandler(x) for x in r.handlers]
 
 
 if __name__ == '__main__':
